requirements.txt
WTForms
Flask
Flask-SQLAlchemy
flask-security-too
Flask-Admin
Flask-Bcrypt
Flask-Mail
python-dotenv
Flask-Migrate
bleach